# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "URLCHEAKER API",
    "description": "SOC-oriented lexical URL triage. The service does not visit submitted URLs.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/analyses": {
      "post": {
        "summary": "Create Analysis",
        "operationId": "create_analysis_api_v1_analyses_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeURLRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Analyses",
        "operationId": "list_analyses_api_v1_analyses_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "classification",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "malicious",
                    "benign",
                    "uncertain"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classification"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 253
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "summary": "Get Analysis",
        "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/verdicts": {
      "post": {
        "summary": "Create Verdict",
        "operationId": "create_verdict_api_v1_analyses__analysis_id__verdicts_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerdictRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerdictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/summary": {
      "get": {
        "summary": "Dashboard Summary",
        "operationId": "dashboard_summary_api_v1_dashboard_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/health": {
      "get": {
        "summary": "System Health",
        "operationId": "system_health_api_v1_system_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "AnalysisListResponse"
      },
      "AnalysisResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "submitted_url": {
            "type": "string",
            "title": "Submitted Url"
          },
          "classification": {
            "type": "string",
            "enum": [
              "malicious",
              "benign",
              "uncertain"
            ],
            "title": "Classification"
          },
          "threat_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Type"
          },
          "malicious_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Malicious Probability"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Confidence"
          },
          "decision_source": {
            "type": "string",
            "enum": [
              "threat_feed",
              "machine_learning",
              "combined"
            ],
            "title": "Decision Source"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonResponse"
            },
            "type": "array",
            "title": "Reasons"
          },
          "feed_matches": {
            "items": {
              "$ref": "#/components/schemas/FeedMatchResponse"
            },
            "type": "array",
            "title": "Feed Matches"
          },
          "requires_analyst_review": {
            "type": "boolean",
            "title": "Requires Analyst Review"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "model_backend": {
            "type": "string",
            "title": "Model Backend"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "submitted_url",
          "classification",
          "threat_type",
          "malicious_probability",
          "confidence",
          "decision_source",
          "reasons",
          "feed_matches",
          "requires_analyst_review",
          "model_version",
          "dataset_version",
          "model_backend",
          "created_at"
        ],
        "title": "AnalysisResponse"
      },
      "AnalyzeURLRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 32768,
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AnalyzeURLRequest"
      },
      "DashboardSummaryResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "malicious": {
            "type": "integer",
            "title": "Malicious"
          },
          "benign": {
            "type": "integer",
            "title": "Benign"
          },
          "uncertain": {
            "type": "integer",
            "title": "Uncertain"
          },
          "requires_review": {
            "type": "integer",
            "title": "Requires Review"
          },
          "threat_feed_decisions": {
            "type": "integer",
            "title": "Threat Feed Decisions"
          }
        },
        "type": "object",
        "required": [
          "total",
          "malicious",
          "benign",
          "uncertain",
          "requires_review",
          "threat_feed_decisions"
        ],
        "title": "DashboardSummaryResponse"
      },
      "FeedMatchResponse": {
        "properties": {
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "threat_type": {
            "type": "string",
            "title": "Threat Type"
          },
          "first_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          }
        },
        "type": "object",
        "required": [
          "source_name",
          "source_record_id",
          "threat_type"
        ],
        "title": "FeedMatchResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status"
          },
          "model_backend": {
            "type": "string",
            "title": "Model Backend"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "feed_records": {
            "type": "integer",
            "title": "Feed Records"
          }
        },
        "type": "object",
        "required": [
          "status",
          "model_backend",
          "model_version",
          "dataset_version",
          "feed_records"
        ],
        "title": "HealthResponse"
      },
      "ReasonResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "feature_name": {
            "type": "string",
            "title": "Feature Name"
          },
          "feature_value": {
            "type": "number",
            "title": "Feature Value"
          }
        },
        "type": "object",
        "required": [
          "code",
          "description",
          "feature_name",
          "feature_value"
        ],
        "title": "ReasonResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerdictRequest": {
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "confirmed_malicious",
              "confirmed_benign",
              "needs_more_analysis",
              "false_positive",
              "false_negative"
            ],
            "title": "Verdict"
          },
          "threat_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Type"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "verdict"
        ],
        "title": "VerdictRequest"
      },
      "VerdictResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict"
          },
          "threat_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Type"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_id",
          "verdict",
          "threat_type",
          "notes",
          "created_at"
        ],
        "title": "VerdictResponse"
      }
    }
  }
}